Sec. 6, App A and B
I added to the main tutoring part in Section 6 what had gone missing,
reworded something in the SY tutoring part, and added one sentence to
the graph description in Appendix A.

I also fixed the ALC appendix, which was a very old version that seems
to be very persistent :) I left the list of new classes as a long
one-under-the-next list, but feel free to turn it into two columns if
that fits better onto the page. I always need to see what comes out
visually. Since I can't do that here, I left the formatting alone.

One mistake I'm not sure I'm qualified to fix is that the appendix B
page 62 reference leads to Testing on page 53, both in the text in
Section 6 on page 50 and in the appendix table of content on page 60. I
found that the reference in the text was "tutoringhours" and in the main
PR file it was tutoringHours. So I changed it in section 6 to
tutoringHours as well. Not sure if it will help.